Crate api/ and move api types into it
